WIP: First steps in refactoring date handling to support -1 for unknown

+++ VRDR/Connectathon.cs
@@ -37,7 +37,7 @@ namespace VRDR
 
             if (record != null && year != null)
             {
-                record.DeathYear = (uint)year;
+                record.DeathYear = year;
             }
 
             return record;
